+++ SignalIntegrity/App/EyeDiagramMeasurementsDialog.py
@@ -19,12 +19,9 @@ EyeDiagramMeasurementsDialog.py
 # If not, see <https://www.gnu.org/licenses/>
 
 import sys
-if sys.version_info.major < 3:
-    import Tkinter as tk
-    from Tkinter import ttk
-else:
-    import tkinter as tk
-    from tkinter import ttk
+
+import tkinter as tk
+from tkinter import ttk
 
 import math
