+++ SignalIntegrity/App/HeaderDialog.py
@@ -18,12 +18,8 @@ HeaderDialog.py
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>
 import sys
-if sys.version_info.major < 3:
-    import Tkinter as tk
-#     import tkFont as font
-else:
-    import tkinter as tk
-#     from tkinter import font
+
+import tkinter as tk
 
 import os
