+++ SignalIntegrity/App/MenuSystemHelpers.py
@@ -19,11 +19,9 @@ MenuSystemHelpers.py
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>
 import sys
-if sys.version_info.major < 3:
-    import Tkinter as tk
-else:
-    import tkinter as tk
-    from tkinter import ttk
+
+import tkinter as tk
+from tkinter import ttk
 
 class ToolTip(object):
     """
