+++ SignalIntegrity/App/PlotWindow.py
@@ -19,10 +19,8 @@ PlotWindow.py
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>
 import sys
-if sys.version_info.major < 3:
-    import Tkinter as tk
-else:
-    import tkinter as tk
+
+import tkinter as tk
     
 import matplotlib
