+++ SignalIntegrity/App/SParameterPropertiesDialog.py
@@ -18,10 +18,8 @@ CalculationPropertiesDialog.py
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>
 import sys
-if sys.version_info.major < 3:
-    import Tkinter as tk
-else:
-    import tkinter as tk
+
+import tkinter as tk
 
 from SignalIntegrity.App.CalculationPropertiesProject import PropertiesDialog,CalculationPropertySI,CalculationProperty
 from SignalIntegrity.Lib.ToSI import nextHigher12458
