+++ SignalIntegrity/App/SParameterViewerWindow.py
@@ -19,12 +19,8 @@ SParameterViewerWindow.py
 # If not, see <https://www.gnu.org/licenses/>
 import sys
 
-if sys.version_info.major < 3:
-    import Tkinter as tk
-    import tkMessageBox as messagebox
-else:
-    import tkinter as tk
-    from tkinter import  messagebox
+import tkinter as tk
+from tkinter import  messagebox
 
 import math
