+++ SignalIntegrity/App/SimulatorDialog.py
@@ -20,12 +20,8 @@ SimulatorDialog.py
 
 import sys
 
-if sys.version_info.major < 3:
-    import Tkinter as tk
-    import tkMessageBox as messagebox
-else:
-    import tkinter as tk
-    from tkinter import messagebox
+import tkinter as tk
+from tkinter import messagebox
 
 from SignalIntegrity.App.SParameterViewerWindow import SParametersDialog
 from SignalIntegrity.App.MenuSystemHelpers import Doer,StatusBar
