+++ Test/CoverageTest.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
-# if coverage doesn't work, try python-coverage
-coverage run --source='../SignalIntegrity/' ./TestSignalIntegrity/TestAll.py > /dev/null
-coverage html -d CoverageReport
-coverage erase
+python3-coverage run --source='../SignalIntegrity/' ./TestSignalIntegrity/TestAll.py > /dev/null
+python3-coverage html -d CoverageReport
+python3-coverage erase
 #read -p "Press [Enter] key to continue..."
 firefox "file:"$PWD"/CoverageReport/index.html" > /dev/null
 read -p "Press [Enter] key to continue..."
