+++ Test/README.md
@@ -5,7 +5,7 @@ Testing of SignalIntegrity is accomplished through the use of the unittest frame
 
 For the most part, the [SignalIntegrityApp](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/SignalIntegrity/App) GUI based application is tested by hand, insofar as the GUI is concerned, but in [TestSignalIntegrity/TestSignalIntegrityApp](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/Test/TestSignalIntegrity/TestSignalIntegrityApp) are tests for testing a _headless_ version of PySIApp that can also be used within a scripted environment.  In other words, the headless version can be utilized to load a project file and obtain whatever results that project file produces.  This is a big aid to scripted applications and [TestSignalIntegrity/TestSignalIntegrityApp](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/Test/TestSignalIntegrity/TestSignalIntegrityApp) tests this headless application.
 
-The [Test](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/Test) directory contains three scripts (for Linux and Windows).  The two Linux scripts are _bash_ scripts and have the extension '.sh'.  In Linux, there is one for Python 2.7 and one for Python 3.  The Windows script is a batch file and has the extension '.bat'.  The script is CoverageTest and is the only test that ever really needs to be run. It runs all of the unit tests for the [SignalIntegrity](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/SignalIntegrity) package.
+The [Test](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/Test) directory contains two scripts (for Linux and Windows).  The Linux scripts is a _bash_ script and has the extension '.sh'.  The Windows script is a batch file and has the extension '.bat'.  The script is CoverageTest and is the only test that ever really needs to be run. It runs all of the unit tests for the [SignalIntegrity](https://github.com/TeledyneLeCroy/SignalIntegrity/tree/master/SignalIntegrity) package.
 
 The script is invoked as either:
 
@@ -13,15 +13,10 @@ The script is invoked as either:
 
 or:
 
-    bash CoverageTest3.sh
-
-depending on which Python version you are using, or:
-
     CoverageTest.bat
 
 depending on your platform.
 
-There is no Python 3 batch file yet.
 
 When the coverage test runs,  it shows something like:
