+++ Test/TestSignalIntegrity/TestBook.py
@@ -22,10 +22,8 @@ from __future__ import print_function
 import unittest
 import os
 import sys
-if sys.version_info.major < 3:
-    from cStringIO import StringIO
-else:
-    from io import StringIO
+
+from io import StringIO
 
 import SignalIntegrity.Lib as si
