+++ Test/TestSignalIntegrity/TestSystemDescription.py
@@ -25,10 +25,7 @@ from numpy import array
 import SignalIntegrity.Lib as si
 
 import sys
-if sys.version_info.major < 3:
-    from cStringIO import StringIO
-else:
-    from io import StringIO
+from io import StringIO
 import os
 
 class TestSystemDescription(unittest.TestCase):
